airplane_ai/services/tts-explain/requirements.txt
fastapi
uvicorn
google-cloud-texttospeech

